backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
pydantic-settings>=2.4.0
svg.path>=6.3
httpx>=0.27.2
python-multipart>=0.0.9
opencv-python-headless>=4.9.0
pillow>=10.0.0
numpy>=1.26.0
openai>=1.50.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
freetype-py>=2.4.0
